requirements.txt
fastapi
pytest
pytest-cov
sqlalchemy
psycopg2
uvicorn
requests
python-dotenv
pyyaml
passlib
PyJWT
